+++ seedling_macros/src/label.rs
@@ -1,7 +1,6 @@
 use bevy_macro_utils::{derive_label, BevyManifest};
 use proc_macro::TokenStream;
 use proc_macro2::TokenStream as TokenStream2;
-use quote::format_ident;
 use quote::quote;
 
 pub fn derive_node_label_inner(input: TokenStream) -> syn::Result<TokenStream2> {
@@ -17,8 +16,8 @@ pub fn derive_node_label_inner(input: TokenStream) -> syn::Result<TokenStream2> 
             type Mutability = #bevy_ecs::component::Immutable;
 
             #[allow(unused_variables)]
-            fn register_component_hooks(hooks: &mut #bevy_ecs::component::ComponentHooks) {
-                hooks.on_insert(|mut world: #bevy_ecs::world::DeferredWorld, context: #bevy_ecs::component::HookContext| {
+            fn on_insert() -> Option<#bevy_ecs::lifecycle::ComponentHook> {
+                Some(|mut world: #bevy_ecs::world::DeferredWorld, context: #bevy_ecs::lifecycle::HookContext| {
                     let value = world.get::<Self>(context.entity).unwrap();
                     let interned = <Self as #label_path>::intern(value);
 
@@ -33,17 +32,12 @@ pub fn derive_node_label_inner(input: TokenStream) -> syn::Result<TokenStream2> 
                         .commands()
                         .entity(context.entity)
                         .insert(labels);
-                });
+                })
             }
         }
     };
 
-    let mut dyn_eq_path = bevy_ecs.clone();
-    dyn_eq_path.segments.push(format_ident!("schedule").into());
-    dyn_eq_path.segments.push(format_ident!("DynEq").into());
-
-    let label_derive: TokenStream2 =
-        derive_label(input, "NodeLabel", &label_path, &dyn_eq_path).into();
+    let label_derive: TokenStream2 = derive_label(input, "NodeLabel", &label_path).into();
 
     Ok(quote! {
         #component_derive
@@ -64,8 +58,8 @@ pub fn derive_pool_label_inner(input: TokenStream) -> syn::Result<TokenStream2> 
             type Mutability = #bevy_ecs::component::Immutable;
 
             #[allow(unused_variables)]
-            fn register_component_hooks(hooks: &mut #bevy_ecs::component::ComponentHooks) {
-                hooks.on_insert(|mut world: #bevy_ecs::world::DeferredWorld, context: #bevy_ecs::component::HookContext| {
+            fn on_insert() -> Option<#bevy_ecs::lifecycle::ComponentHook> {
+                Some(|mut world: #bevy_ecs::world::DeferredWorld, context: #bevy_ecs::lifecycle::HookContext| {
                     let value = world.get::<Self>(context.entity).unwrap();
                     let container = ::bevy_seedling::pool::label::PoolLabelContainer::new(value, context.component_id);
 
@@ -73,17 +67,12 @@ pub fn derive_pool_label_inner(input: TokenStream) -> syn::Result<TokenStream2> 
                         .commands()
                         .entity(context.entity)
                         .insert(container);
-                });
+                })
             }
         }
     };
 
-    let mut dyn_eq_path = bevy_ecs.clone();
-    dyn_eq_path.segments.push(format_ident!("schedule").into());
-    dyn_eq_path.segments.push(format_ident!("DynEq").into());
-
-    let label_derive: TokenStream2 =
-        derive_label(input, "PoolLabel", &label_path, &dyn_eq_path).into();
+    let label_derive: TokenStream2 = derive_label(input, "PoolLabel", &label_path).into();
 
     Ok(quote! {
         #component_derive