+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.6.0-rc.1
+
+## Fixes
+
+- `PlaybackSettings::play` no longer restarts from the beginning
+- Fixed missing `EffectId` for simple nodes
+- Fixed dangling sample players when their sampler is stolen
+- Improved robustness of delay line indexing
+
 # 0.5.2
 
 ## Fixes